+++ src/2-two/postTagger/model/verbs/imperative.js
@@ -47,4 +47,10 @@ export default [
   { match: '^do not [#Infinitive]', group: 0, tag: 'Imperative', reason: 'do-not-be' },
   // allow yourself
   { match: '[#Infinitive] (yourself|yourselves)', group: 0, tag: 'Imperative', reason: 'allow-yourself' },
+  // continue playing
+  { match: '^[#Infinitive] #Gerund', group: 0, tag: 'Imperative', reason: 'keep-playing' },
+  // go to it
+  { match: '^[#Infinitive] (to|for|into|toward)', group: 0, tag: 'Imperative', reason: 'go-to' },
+  // relax and unwind
+  { match: '^[#Infinitive] (and|or) #Infinitive', group: 0, tag: 'Imperative', reason: 'inf-and-inf' },
 ]