+++ changelog.md
@@ -15,6 +15,7 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 - **[change]** - fixes to .clauses() parser
 - **[change]** - an astrix is not a word
 - **[new]** - @hasColon method
+- **[new]** - @hasDash supports two dashes
 
 #### 14.5.2 [Oct 2022]
 - **[fix]** - runtime error #965